requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.10.0
dnspython==2.1.0
email-validator==1.1.2
essential-generators==1.0
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
langdetect==1.0.8
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
nltk==3.5
numpy==1.20.1
oauthlib==3.0.1
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.1.2
pycparser==2.20
pyOpenSSL==19.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
qrcode==6.1
regex==2021.3.17
requests==2.21.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
tqdm==4.57.0
tweepy==3.10.0
urllib3==1.24.3
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==2.0.1
